refactor(ar.io cli): derive more types from AoMethodParams PE-5854
